requirements.txt
extract-msg>=0.41.0  # Using latest stable version
tenacity>=8.2.0
# boto3>=1.28.0
python-dotenv>=1.0.0  # For environment variable management
pandas
pyyaml
opencv-python  # For drawing bounding boxes
pdf2image  # For converting PDFs to images
pymupdf
pypdf
streamlit>=1.28.0
numpy
azure-ai-documentintelligence
azure-core
azure-storage-blob
openai>=1.3.0
pyodbc
openpyxl
tabulate
scikit-learn
matplotlib      # Optional - for visualization
seaborn
rapidfuzz
uvicorn
uvicorn[standard]
fastapi
python-multipart
pydantic
xlsxwriter
opencensus-ext-azure
opencensus-ext-logging
requests>=2.31.0
email-validator==2.3.0
